requirements.txt
aiohttp
beautifulsoup4
aiogram
python-dotenv